mcp_server/requirements.txt
fastmcp>=0.9.0
uvicorn[standard]>=0.29.0
starlette>=0.37.0

# Auth
pyjwt>=2.8.0
supabase>=2.4.0

# Vector DB
pinecone>=3.0.0
openai>=1.30.0           # for embedding generation (swap if using another provider)

# Graph DB
neo4j>=5.19.0

# Config
pydantic-settings>=2.2.0
python-dotenv>=1.0.0